subsystem: Test refactoring
